messages: store message in global data file
This makes it easier to add/remove messages.

This commit is part of an effort to increase ease of use and maintenance
for future parkour club committees, who may be non-technical and will
not have the level of familiarity I have built up with this system.

Future commits may allow messages to be written in markdown instead.